requirements.txt
aiofiles==24.1.0
aiogram==3.22.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
APScheduler==3.11.0
asttokens==3.0.0
async-timeout==5.0.1
attrs==25.3.0
certifi==2025.8.3
charset-normalizer==3.4.3
decorator==5.2.1
et_xmlfile==2.0.0
exceptiongroup==1.3.0
executing==2.2.1
frozenlist==1.7.0
idna==3.10
ipython==8.37.0
jedi==0.19.2
magic-filter==1.0.12
matplotlib-inline==0.1.7
multidict==6.6.4
numpy==2.2.6
openpyxl==3.1.5
parso==0.8.5
pexpect==4.9.0
prompt_toolkit==3.0.52
propcache==0.3.2
ptyprocess==0.7.0
pure_eval==0.2.3
pydantic==2.11.9
pydantic_core==2.33.2
Pygments==2.19.2
python-dotenv==1.1.1
requests==2.32.5
stack-data==0.6.3
traitlets==5.14.3
typing-inspection==0.4.1
typing_extensions==4.15.0
tzlocal==5.3.1
urllib3==2.5.0
wcwidth==0.2.14
yarl==1.20.1
